+++ src/main/java/com/solvd/tickets/source/jira/JiraTicket.java
@@ -15,7 +15,35 @@
  * SOFTWARE.
  */
 
+package com.solvd.tickets.source.jira;
+
+import com.solvd.tickets.tickets.Ticket;
+import lombok.RequiredArgsConstructor;
+
 /**
- * Models.
+ * Jira ticket.
  */
-package com.solvd.tickets.model;
+@RequiredArgsConstructor
+public class JiraTicket implements Ticket {
+
+  /**
+   * Id.
+   */
+  private final String id;
+
+  /**
+   * Issue.
+   */
+  private final String summary;
+
+  @Override
+  public String id() {
+    return this.id;
+  }
+
+  @Override
+  public String summary() {
+    return this.summary;
+  }
+
+}
